feat(game): Implements comprehensive in-game justice and trial system (#1453)
* feat(game): Implement comprehensive in-game justice and trial system

This commit introduces a fully-fledged justice system, allowing for:
- Criminal reporting and evidence generation, tracked in the `CrimeManager`.
- Arrest and trial initiation for wanted players.
- Courtroom settings and jail locations configurable via `Justice.json`.
- Player participation as defendants, jury members, or audience.
- Dedicated chat channels for courtroom proceedings.
- Integration with existing game mechanics (portals, dungeons, combat, character buffs).
- Updated character stats for crime, infamy, and jury points.
- New client-server packets implementation for trial UI and interactions.
- GM commands for testing and administration of the justice system.
- Database schema updates for crime events to include judgement details.
- Fixes a bug in ChatManager initialization where the Judge channel did not have the correction faction Ids assigned.

* feat(game): More chat related fixes

* feat(game): Enhance justice system with pirate mechanics

This commit refines the in-game justice and trial system, particularly improving pirate character interactions and overall trial flow.

- **Pirate Justice**: Implements conditional arrests for pirates, allowing them only in non-desperado zones. Ensures trials proceed by defaulting to the arrestor's court region if the criminal is a pirate.
Manages the `Contemptuous` buff to identify pirate characters.
- **Trial Flow**: Adds a dedicated chat type for trial participants. Improves logic for finding active trial chat channels. Prevents players already involved in a trial from queuing for jury duty. Ensures defendants are alive at the start of a trial and removes selected jury members from queues.

* feat(game): Refine crime types

- Add new `CrimeKind` entries, `BotReport` and `FalseBotReport`.
- Renamed `CrimeKind.None` to `CrimeKind.Invalid` and `CrimeKind.Unknown1` to `CrimeKind.None` to follow the naming of more recent naming convention discoveries.

* feat(game): bounty board packets

Changed it so the bounty boards will now just open with a empty board.

* feat(game): WIP Bot reports

WIP of bot report functions

* Tradepack debug stuff

* feat(game): Implement offline trial processing and enhance faction changes

- Introduce logic to process trials for characters found guilty while offline, automatically sending them to court jail upon login.
- Allow pirate characters to revert to their original faction if their infamy points drop to zero or below.
- Refactor character-related counters, including justice statistics, into a dedicated partial class (`CharacterCounters.cs`) for better organization.
- Improve the `SetFaction` mechanism to handle related game elements like housing, expeditions, and parties more robustly when a character's faction changes.
- Add new database columns to persist justice counters and offline guilty status.

* feat(game): Enhance bot reporting and trial outcome tracking

- Introduce `reported_as_bot_count` to track how many times a character is reported as a bot.
- Centralize and refine the update logic for various trial-related counters, including guilty, not guilty, and accepted plea statuses.
- Add logging for bot reporting activities, including when a reported bot reaches the prime suspect threshold.
- Fix a typo in method documentation.

* feat(game): Enable custom messages for bot reports and clarify skill object types

- The `ReportBot` function now accepts a custom message from the reporter, providing valuable context for investigations.
- Generic `SkillObjectUnk` types have been renamed to more descriptive names to improve code clarity and maintainability.

* -feat(game): greptile suggestions

- Prevent players from canceling trials they are not defendants in.
- Validate jury verdicts to ensure they originate from the assigned jury member.
- Correctly initialize all jury seats within a courtroom.
- Enhance thread safety for reported suspect tracking using `ConcurrentDictionary`.
- Fix an incorrect multiplier for jail time duration calculations.

* feat(game): Fixed trial sentence durations for guilty verdicts

Put the multipliers for final verdict times more in line with what the client suggests.

+++ AAEmu.Game/Models/Game/Chat/ChatChannel.cs
@@ -30,7 +30,7 @@ public class ChatChannel
     /// <summary>
     /// Internal Id
     /// </summary>
-    public long InternalId { get; init; }
+    public uint InternalId { get; init; }
 
     /// <summary>
     /// Internal name of this channel
